+++ qlab/problems/__init__.py
@@ -8,8 +8,9 @@ from qlab.problems import (  # noqa: F401  (import = registration)
     grover,
     maxcut,
     qft,
+    qpe,
     simon,
     state_prep,
 )
 
-__all__ = ["state_prep", "maxcut", "bernstein_vazirani", "deutsch_jozsa", "simon", "grover", "qft"]
+__all__ = ["state_prep", "maxcut", "bernstein_vazirani", "deutsch_jozsa", "simon", "grover", "qft", "qpe"]